tests, net: Add a bridge network-attachment-definition helper
Create a bridge net-attach-def helper to share and uniform the usage of
this common resource.

The plugin name has been changed from `cnv-bridge` to `bridge`, as
`cnv-bridge` is an alias (more accurately, a symlink) to `bridge`.
In the past the two were not identical, but that time is no more and it
is considered a leftover.

Signed-off-by: Edward Haas <edwardh@redhat.com>