tests, net, sriov: Drop IPAM usage in SR-IOV net-attach-def
Kubevirt uses SR-IOV in VFIO mode, i.e. the VF devices are not using a
network driver on the host (or pod) and therefore do not show as a
network interface. As such, these devices cannot be assigned any IP
address in the pod, making the CNI IPAM plugin irrelevant.

Therefore, the SR-IOV network-attachment-definition should not have any
ref to IPAM. This change removes such references.

Signed-off-by: Edward Haas <edwardh@redhat.com>