cleanup(virtctl): Use context to pass client config
Use the pkg/virtctl/clientconfig module to store and retrieve the client
config from/in the context used by virtctl commands. This allows to stop
passing the client config generated in NewVirtctlCommand down to every
command that is registered in the root command and also to stop
returning the client config from the function.

Commands that have flags retain their command struct, while commands
which do not have any flags are refactored to drop empty command
structs. In case the command struct is not already used to store values
from flags this needs to be changed in a follow up.

Signed-off-by: Felix Matouschek <fmatouschek@redhat.com>