tests: verify volume migration with the persistent state
The TPM adds an additional PVCs to the virt-launcher pod in order to
persist the TPM state. This new volumes needs to be excluded by the
volume migration and the VM needs to be storage-migratable even if the
TPM is stored on a RWO volume.

This change enables the feature for the persistent state and configure
the RWO storage class for the backend storage.

Signed-off-by: Alice Frosi <afrosi@redhat.com>