repeated_test.sh: do not run test with PERIODIC
Since we have added the marker PERIODIC to a test for a reason, namely
that it causes troubles when being run in any presubmit lane, we exclude
it also from running inside the check-test-for-flakes-lane.

The presence of the decorator indicates that it is a work item, also the
signal from the periodic test lanes every now and then is enough to
remind us that we need to make a test more resilient.

Signed-off-by: Daniel Hiller <dhiller@redhat.com>